refactor: restructure Onboarding module
Onboarding now has two subpackages: api & impl. Code is currently in impl.
