+++ feature/onboarding/impl/src/main/kotlin/com/potatosheep/kite/feature/onboarding/impl/nav/OnboardingNavigation.kt
@@ -1,11 +1,11 @@
-package com.potatosheep.kite.feature.onboarding.nav
+package com.potatosheep.kite.feature.onboarding.impl.nav
 
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
-import com.potatosheep.kite.feature.onboarding.OnboardingRoute
+import com.potatosheep.kite.feature.onboarding.impl.OnboardingRoute
 import kotlinx.serialization.Serializable
 
 @Serializable
