+++ feature/onboarding/impl/src/main/res/values/colors.xml
