+++ feature/onboarding/impl/src/main/res/values/strings.xml
