+++ settings.gradle.kts
@@ -53,7 +53,8 @@ include(":core:database")
 include(":feature:bookmark:impl")
 include(":feature:bookmark:api")
 include(":feature:exception")
-include(":feature:onboarding")
+include(":feature:onboarding:impl")
+include(":feature:onboarding:api")
 include(":feature:about:impl")
 include(":feature:about:api")
 include(":core:media")
